requirements.txt
openai
langchain
faiss-cpu
flask
python-dotenv
PyPDF2
tiktoken
pypdf